refactor(xmarkico): place xmark icon into components dir
n
